Move serializer to utils namespace
#12
